% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pois_eit}
\alias{pois_eit}
\title{pois_eit(lambda, size)}
\usage{
pois_eit(lambda, size)
}
\arguments{
\item{lambda}{expected value}

\item{size}{number of elements}
}
\value{
a vector of Poisson integers
}
\description{
Generate a vector of Poisson random values based upon exponential
  inter-arrival times (Devroye).
}
\details{
Works for large values of lambda.
}
\examples{
pois_eit(1, 10)

}
