% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pois_iss}
\alias{pois_iss}
\title{pois_iss(lambda, size)}
\usage{
pois_iss(lambda, size)
}
\arguments{
\item{lambda}{expected value}

\item{size}{number of elements}
}
\value{
a vector of Poisson integers
}
\description{
Generate a vector of Poisson random values using inversion by sequential
  search (Devroye).
}
\details{
Fails to terminate for lambda greater than 102.873.
}
\examples{
pois_iss(1, 10)

}
