% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pois_knuth}
\alias{pois_knuth}
\title{pois_knuth(lambda, size)}
\usage{
pois_knuth(lambda, size)
}
\arguments{
\item{lambda}{expected value}

\item{size}{number of elements}
}
\value{
a vector of Poisson integers
}
\description{
Generate a vector of Poisson random values (Knuth).
}
\details{
For each value generated, algorithmic complexity is proportional to
  parameter lambda. Underflow occurs for lambda greater than 103.972,
  resulting in junk data.
}
\examples{
pois_knuth(1, 10)

}
