% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pois_mult}
\alias{pois_mult}
\title{pois_mult(lambda, size)}
\usage{
pois_mult(lambda, size)
}
\arguments{
\item{lambda}{expected value}

\item{size}{number of elements}
}
\value{
a vector of Poisson integers
}
\description{
Generate a vector of Poisson random values using multiplication
  of uniform random variates (Devroye).
}
\details{
Underflows for lambda greater than 103.972, resulting in junk data.
}
\examples{
pois_mult(1, 10)

}
