% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppPkgDemo.R
\docType{package}
\name{RcppPkgDemo}
\alias{RcppPkgDemo}
\alias{RcppPkgDemo-package}
\title{RcppPkgDemo: An Example of Creating an R Package with Functions Implemented in C++ using Rcpp}
\description{
This project uses Rcpp to create an R package implemented in C++.
    I use Roxygen2 to manage documentation, which turns specially formatted
    comments into .Rd files. These are the help files displayed when you
    run ?foo() in the R console, where foo() is a function.
    The package consists of C++ functions for generating random numbers from the Poisson
    distribution using algorithms described in the book Non-Uniform Random Variate Generation
    by Luc Devroye.
    Random number generators for various distributions are needed for many
    applications, such as sampling or modeling stochastic processes.
    This project is intended as an exercise in package development with C++ for R.
    It is also a good reference and reusable foundation, since the options and setup
    reflect the type of requirements I'll most likely encounter in future projects.
}
