% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumI}
\alias{sumI}
\title{sumI}
\usage{
sumI(x)
}
\arguments{
\item{x}{vector to sum}
}
\value{
The sum of all elements of x as a double
}
\description{
Sum vector elements using C++ with an iterator
}
\examples{
x <- runif(1e3)
print(sumI(x))

}
