% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumC}
\alias{sumC}
\title{sumC
Sum vector elements using C++}
\usage{
sumC(x)
}
\arguments{
\item{x}{vector to sum}
}
\value{
The sum of all elements of x as a double
}
\description{
sumC
Sum vector elements using C++
}
\examples{
x <- runif(1e3)
print(sumC(x))

}
