% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumI}
\alias{sumI}
\title{Sum NumericVector elements using C++ with an iterator}
\usage{
sumI(x)
}
\arguments{
\item{x}{NumericVector to sum}
}
\value{
The sum of all elements of x as a double
}
\description{
Sum NumericVector elements using C++ with an iterator
}
\examples{
\dontrun{
x <- runif(1e3)
print(sumI(x))
}

}
